+++ src/com/intellij/idea/plugin/hybris/system/cockpitng/codeInsight/completion/CngCompletionContributor.kt
@@ -38,9 +38,14 @@ class CngCompletionContributor : CompletionContributor() {
         )
         extend(
             CompletionType.BASIC,
-            PlatformPatterns.psiElement().inside(CngPatterns.FLOW_STEP_CONTENT_PROPERTY_QUALIFIER),
-            CngFlowPropertyQualifierCompletionProvider.instance
+            PlatformPatterns.psiElement().inside(CngPatterns.FLOW_STEP_CONTENT_PROPERTY_LIST_PROPERTY_QUALIFIER),
+            CngFlowPropertyListPropertyQualifierCompletionProvider.instance
         )
+//        extend(
+//            CompletionType.BASIC,
+//            PlatformPatterns.psiElement().inside(CngPatterns.FLOW_STEP_CONTENT_PROPERTY_QUALIFIER),
+//            CngFlowPropertyQualifierCompletionProvider.instance
+//        )
         extend(
             CompletionType.BASIC,
             PlatformPatterns.psiElement().inside(CngPatterns.FLOW_INITIALIZE_TYPE),