+++ src/com/intellij/idea/plugin/hybris/system/cockpitng/psi/reference/CngFlowTSItemAttributeReference.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -20,9 +20,13 @@ package com.intellij.idea.plugin.hybris.system.cockpitng.psi.reference
 
 import com.intellij.idea.plugin.hybris.system.cockpitng.psi.CngPsiHelper
 import com.intellij.idea.plugin.hybris.system.type.psi.reference.AbstractAttributeDeclarationReference
+import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 
-class CngFlowTSItemAttributeReference(element: PsiElement) : AbstractAttributeDeclarationReference(element) {
+class CngFlowTSItemAttributeReference : AbstractAttributeDeclarationReference {
+
+    constructor(element: PsiElement) : super(element)
+    constructor(element: PsiElement, textRange: TextRange) : super(element, textRange)
 
     override fun resolveType(element: PsiElement) = CngPsiHelper.resolveContextTypeForNewItemInWizardFlow(element)